Добавлен кошелёк, форматирование валюты

+++ Assets/_Project/Scripts/Core/Interface/ICollectable.cs
@@ -1,4 +1,4 @@
 public interface ICollectable
 {
-    bool TryCollect(out long value);
+    bool TryCollect(out float value);
 }
